Home

FTPServer.jl

FTPServer.Server

FTPServer.cleanup

FTPServer.init

FTPServer.serve

FTPServer.Server — Type.

Server(homedir=/home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/data; username=user, password=12345, permissions=elradfmwM, security=:none)

A Server stores settings for create an pyftpdlib server.

Arguments

homedir::AbstractString: Directory where you want store to store your data for the test server.

Keywords

username: Default login username

password: Default login password

permission: Default user read/write permissions

security: Security method to use for connecting (options: :none, :implicity, :explicity). Passing in :none will use FTP and passing in :implicity or :explicity will use the appropriate FTPS connection.

source

FTPServer.cleanup — Method.

cleanup()

Cleans up the default FTPServer.ROOT directory:

/home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/data

/home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/test.crt

/home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/test.key

source

FTPServer.init — Method.

init()

Creates a test /home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/data with a few sample files if one hasn't already been setup.

/home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/data/test_download.txt /home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/data/test_download2.txt /home/travis/build/invenia/FTPServer.jl/deps/usr/ftp/data/test_directory/`

source

FTPServer.serve — Method.

serve(f, args...; kwargs...)

Passes args and kwargs to the Server constructor and runs the function f by passing in the server instance. Upon completion the server will automatically be shutdown.

source